# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A modern RESTful API for managing travel plans, including trips, destinations, accommodations, places, and restaurants.",
        "title": "Travel Planning API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support Team",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/insert_csv/destination": {
            "post": {
                "description": "Import destination data from a CSV file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "InsertData"
                ],
                "summary": "Insert Destination Data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/insert_csv/hotel": {
            "post": {
                "description": "Import hotel data from a CSV file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "InsertData"
                ],
                "summary": "Insert Hotel Data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/insert_csv/place": {
            "post": {
                "description": "Import place data from a CSV file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "InsertData"
                ],
                "summary": "Insert Place Data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/insert_csv/restaurant": {
            "post": {
                "description": "Import restaurant data from a CSV file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "InsertData"
                ],
                "summary": "Insert Restaurant Data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/trip/create": {
            "post": {
                "description": "Create a new trip with destinations, accommodations, activities, and restaurants",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trip"
                ],
                "summary": "Create a new trip",
                "parameters": [
                    {
                        "description": "Trip Details",
                        "name": "trip",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTripRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns trip ID",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/trip/get_plan": {
            "post": {
                "description": "Get AI-generated trip suggestions based on user preferences",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trip"
                ],
                "summary": "Get trip suggestions",
                "parameters": [
                    {
                        "description": "Trip Suggestion Parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TripSuggestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suggested trip",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CreateTripRequestByDate"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/trip/save": {
            "put": {
                "description": "Update an existing trip with new details including destinations, accommodations, activities, and restaurants",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trip"
                ],
                "summary": "Save changes to an existing trip",
                "parameters": [
                    {
                        "description": "Trip Details",
                        "name": "trip",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TripDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trip updated successfully",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/trip/{id}": {
            "get": {
                "description": "Get complete trip details including destinations, accommodations, activities, and restaurants",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trip"
                ],
                "summary": "Get trip details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trip details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TripDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid trip ID",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "404": {
                        "description": "Trip not found",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AccommodationSuggestion": {
            "type": "object",
            "properties": {
                "accommodation_id": {
                    "type": "string"
                },
                "booking_link": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "destination_id": {
                    "type": "string"
                },
                "elderly_friendly": {
                    "type": "boolean"
                },
                "image_url": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Image"
                    }
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "rating": {
                    "type": "number"
                },
                "room_info": {
                    "type": "string"
                },
                "room_types": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RoomType"
                    }
                },
                "tax_info": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "dto.AccommodationsSuggestion": {
            "type": "object",
            "properties": {
                "accommodations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AccommodationSuggestion"
                    }
                }
            }
        },
...
        "dto.CreateTripAccommodationRequest": {
            "type": "object",
            "properties": {
                "accommodation_id": {
                    "type": "string"
                },
                "check_in_date": {
                    "type": "string"
                },
                "check_out_date": {
                    "type": "string"
                },
                "cost": {
                    "type": "number"
                },
                "notes": {
                    "type": "string"
                },
                "trip_destination_id": {
                    "type": "string"
                }
            }
        },
        "dto.CreateTripDestinationRequest": {
            "type": "object",
            "properties": {
                "accommodations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CreateTripAccommodationRequest"
                    }
                },
                "arrival_date": {
                    "type": "string"
                },
                "departure_date": {
                    "type": "string"
                },
                "destination_id": {
                    "type": "string"
                },
                "order_num": {
                    "type": "integer"
                },
                "places": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CreateTripPlaceRequest"
                    }
                },
                "restaurants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CreateTripRestaurantRequest"
                    }
                },
                "trip_id": {
                    "type": "string"
                }
            }
        },
        "dto.CreateTripPlaceRequest": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "place_id": {
                    "type": "string"
                },
                "scheduled_date": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "trip_destination_id": {
                    "type": "string"
                }
            }
        },
        "dto.CreateTripRequest": {
            "type": "object",
            "properties": {
                "budget": {
                    "type": "number"
                },
                "end_date": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "trip_destinations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CreateTripDestinationRequest"
                    }
                },
                "trip_name": {
                    "type": "string"
                },
                "trip_status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "dto.CreateTripRequestByDate": {
            "type": "object",
            "properties": {
                "destination": {
                    "type": "string"
                },
                "end_date": {
                    "description": "\"YYYY-MM-DD\"",
                    "type": "string"
                },
                "plan_by_day": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PlanByDay"
                    }
                },
                "start_date": {
                    "description": "\"YYYY-MM-DD\"",
                    "type": "string"
                },
                "trip_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "dto.CreateTripRestaurantRequest": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "meal_date": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "reservation_info": {
                    "type": "string"
                },
                "restaurant_id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "trip_destination_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.PlaceSuggestion": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "categories": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "destination_id": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Image"
                    }
                },
                "main_image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "opening_hours": {
                    "type": "string"
                },
                "place_id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "rating": {
                    "type": "number"
                },
                "reviews": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "dto.PlacesSuggestion": {
            "type": "object",
            "properties": {
                "places": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PlaceSuggestion"
                    }
                }
            }
        },
        "dto.PlanByDay": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "\"YYYY-MM-DD\"",
                    "type": "string"
                },
                "day_title": {
                    "description": "e.g., \"Ngày 1: Check-in \u0026 khám phá biển\"",
                    "type": "string"
                },
                "segments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Segment"
                    }
                }
            }
        },
...
        "dto.RestaurantsSuggestion": {
            "type": "object",
            "properties": {
                "restaurants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.RestaurantSuggestion"
                    }
                }
            }
        },
        "dto.Segment": {
            "type": "object",
            "properties": {
                "activities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Activity"
                    }
                },
                "time_of_day": {
                    "description": "e.g., \"morning\", \"afternoon\", \"evening\"",
                    "type": "string"
                }
            }
        },
...
        "dto.TripDTO": {
            "type": "object",
            "properties": {
                "budget": {
                    "type": "number"
                },
                "end_date": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "trip_destinations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TripDestinationDTO"
                    }
                },
                "trip_id": {
                    "type": "string"
                },
                "trip_name": {
                    "type": "string"
                },
                "trip_status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "dto.TripDestinationDTO": {
            "type": "object",
            "properties": {
                "accommodations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TripAccommodationDTO"
                    }
                },
                "arrival_date": {
                    "type": "string"
                },
                "departure_date": {
                    "type": "string"
                },
                "destination_id": {
                    "type": "string"
                },
                "order_num": {
                    "type": "integer"
                },
                "places": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TripPlaceDTO"
                    }
                },
                "restaurants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TripRestaurantDTO"
                    }
                },
                "trip_destination_id": {
                    "type": "string"
                },
                "trip_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.TripSuggestionRequest": {
            "type": "object",
            "properties": {
                "accommodation": {
                    "$ref": "#/definitions/dto.AccommodationsSuggestion"
                },
                "places": {
                    "$ref": "#/definitions/dto.PlacesSuggestion"
                },
                "restaurants": {
                    "$ref": "#/definitions/dto.RestaurantsSuggestion"
                }
            }
        },
...
        "model.Image": {
            "type": "object",
            "properties": {
                "alt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "model.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Enter the token with the `Bearer: ` prefix, e.g. \"Bearer abcde12345\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}